Implement `insert` and `clean` test helpers
These helpers are useful for writing DB tests. `insert` provides
a convenient way to add data for test and `clean` removes all
data from the DB so you can clean up between tests.

There is also a change here to build the code at the top level,
this is for convenience of requiring the relevant modules when
using this as a node module.